Services interfaces Add interfaces for all services. Moved all services to portfolio.Services namespace. Rename migrations because my IDE was trying to use the Admin migration instead of the Admin model. Did some other minor refactoring.
