+++ portfolio/Controllers/PersonalInfoController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using portfolio.Models;
 using portfolio.Repositories;
-using portfolio.Storage;
+using portfolio.Services;
 
 namespace portfolio.Controllers;
