+++ portfolio/Interfaces/ServicesInterfaces/EmailInterface.cs
@@ -1,0 +1,6 @@
+namespace portfolio.Interfaces;
+
+public interface IEmailService
+{
+    Task SendEmailAsync(string email, string subject, string message);
+}
