+++ portfolio/Interfaces/ServicesInterfaces/ImageInterface.cs
@@ -1,0 +1,6 @@
+namespace portfolio.Interfaces;
+
+interface IStorageService
+{
+    Task<string> UploadFileAsync(Stream fileStream, string fileName);
+}
